ghapp/requirements.txt
click
attrs
cattrs
aiohttp
aiorun
frozendict
giturlparse
decorator
